// desafio-frontend/src/Types/Users.ts
import { Employment } from './Employment';
import { Address } from './Address';
import { CreditCard } from './CreditCard';
import { Subscription } from './Subscription';

export interface User {
  id: number;
  uid: string;
  password: string;
  first_name: string;
  last_name: string;
  username: string;
  email: string;
  avatar: string;
  gender: string;
  phone_number: string;
  social_insurance_number: string;
  date_of_birth: string;
  employment: Employment;
  address: Address;
  credit_card: CreditCard;
  subscription: Subscription;
}